SCRUM-126 Task 4.1: Configure EF Core tools and dev connection strings (LocalDB + Docker SQL)
