+++ AIMS/Legacy/Scaffolded/Dtos/AddHardwareDto.cs
