+++ AIMS/Legacy/Scaffolded/Models/Hardware.cs
