+++ AIMS/Program.cs
@@ -10,14 +10,20 @@ builder.Services.AddSwaggerGen(); // take out after development
 
 
 builder.Services.AddDbContext<AimsDbContext>(options =>
-    options.UseSqlServer(builder.Configuration.GetConnectionString("DefaultConnection")));
+{
+    var cs = builder.Configuration.GetConnectionString(
+        builder.Environment.IsDevelopment() &&
+        Environment.OSVersion.Platform == PlatformID.Win32NT
+            ? "DefaultConnection"   // LocalDB on Windows
+            : "DockerConnection"    // Docker SQL on Mac/Linux
+    );
+    options.UseSqlServer(cs);
+});
 
 var app = builder.Build();
 
-
 app.UseSwagger(); // take out after development
 app.UseSwaggerUI(); //used for testing APIs. Swagger UI will be available at /swagger/index.html
-
 
 // Configure the HTTP request pipeline.
 if (!app.Environment.IsDevelopment())
@@ -39,7 +45,4 @@ app.MapControllerRoute(
     pattern: "{controller=Home}/{action=Index}/{id?}")
     .WithStaticAssets();
 
-
 app.Run();
-
-
