+++ AIMS/appsettings.Development.json
@@ -4,5 +4,9 @@
       "Default": "Information",
       "Microsoft.AspNetCore": "Warning"
     }
+  },
+  "ConnectionStrings": {
+    "DefaultConnection": "Server=(localdb)\\MSSQLLocalDB;Database=AIMS;Trusted_Connection=True;MultipleActiveResultSets=true",
+    "DockerConnection": "Server=localhost,1433;Database=AIMS;User Id=sa;Password=StrongP@ssword!;TrustServerCertificate=True;"
   }
 }
